Fixes #67 Serialize access for a user so X-L-M is always unique
This change ensures that X-Last-Modified is monotonically increases
whenever there is a change (POST, PUT, DELETE) to the user's data. This
is accomplished simply by delaying a request by 10ms if the data was
changed < 10ms ago.

Also a *very* huge bug was found while adding this feature. The previous
db.DeleteBSOs() function did not include the collection Id so it would
delete all BSOs with the same Id regardless of the collection it was in.