Fix broken logging of token payload in PR #137 (#138)
In PR #137 the hawkHandler creates a *new* request with the token data
in the session context. Since the logHandler wraps all other handlers
this means the logHandler's request object's context is not the same as
the one passed down through the HawkHandler.

The solution is to be smarter with how the session is handled. The
logHandler will reuse or create a session before passing the request
along. The HawkHandler will also do this so it does not rely on any
upstream handler's to reduce coupling between handlers.

A new test was added to make sure a session is correctly reused so the
fxa_uid and device_id can be logged.