Fix bug 1332552. Treat bsos with ttl:null as never expiring. (#157)
- Fix bug 1332552. Treat bsos with ttl:null as never expiring.

The 1.5 spec says that undefined and ttl=null should mean the BSO never expires.
With this commit, when ttl:null is encounted it will be changed into 100
years (in seconds). The default BSO TTL is no 100 years as well.

However there now there is a subtle difference between undefined
TTL and ttl=null between insert/update behaviour:

TTL is undefined (not in the JSON)

    PUT .../storage/col/bsoId
     - insert: TTL is set to 100 years
     - update: TTL is *not* updated

TTL=null

    PUT .../storage/col/bsoId
     - insert: TTL is set to 100 years
     - update: TTL is set to 100 years

-  Add invalid data tests for web.parseIntoBSO. Fix bug where id=null is ok

- Remove usage of PYTHON_TEST_CONTAINER in circle.yml
In practice having the ENV var and the hard coded value means you have to change it in two places. Better to have it hard coded and update it with a git record.