Fix #169 so gosync has the proper DELETE collection behaviour (#171)
The logic in db.DeleteCollection was updated to be correct. When a
collection is deleted its last modified timestamp is set to 0 and the
storage level last modified timestamp is set to the current time.

The last modified timestamp at the storage level is now discrete from
the max(last-modified) of collections. It was a wrong assumption that
the two values were always the same.

Unit tests were added to ensure the change in functionality is correct.